Create User model with route & repository

+++ conf/db/migration/auth_service/V1__create_initial_tables.sql
@@ -1,4 +1,4 @@
 CREATE TABLE users (
-    id integer PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name varchar(40)
 );
